+++ frontend/src/components/chatwindow/Chatwindow.tsx
@@ -119,7 +119,7 @@ const Chatwindow: React.FC<{ $display: boolean, z?: number }> = ({
         return response.json();
       })
       .then((res: string[]) => setTabs(res))
-      .catch((error) => console.log(error));
+      .catch((error) => console.error(error));
     setRoom(tabs[tabs.length - 1]);
   }, []); // eslint-disable-line react-hooks/exhaustive-deps
